requirements.txt
typing~=3.7.4.3
Pillow~=9.1.1